components.del/organaizer.del/pyproject.toml
[tool.poetry]
name = "organaizer"
version = "0.1.0"
description = ""
authors = ["Gabriel Gutierrez <gutierrez.ge@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pydantic = "^2.9.2"
python-dotenv = "^1.0.1"
minio = "^7.2.13"
sqlalchemy = "^2.0.36"
psycopg2-binary = "^2.9.10"
ultralytics = "^8.3.55"
opencv-python = "^4.10.0.84"
py3dbp = "^1.1.2"
fastapi = "^0.115.6"
uvicorn = "^0.34.0"
pydantic-settings = "^2.7.1"
kivy = "^2.3.1"
kivymd = "^1.2.0"

[tool.poetry.dev-dependencies]
pytest = "^8.3.3"
pytest-mock = "^3.14.0"
moto = "^5.0.18"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
